moved cxvalues to its own file
